# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TransitOps API",
    "description": "Fleet operations platform \u2014 manage vehicles, drivers, trips, maintenance, fuel, expenses, and reports.",
    "version": "1.0.0"
  },
  "paths": {
    "/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Register a new user",
        "description": "Create a new user account. No authentication required.",
        "operationId": "register_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRegister"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "409": {
            "description": "Email already registered"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Login and obtain JWT",
        "description": "Authenticate with email and password. Returns a JWT access token.",
        "operationId": "login_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLogin"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Get current user profile",
        "description": "Returns the profile of the currently authenticated user.",
        "operationId": "get_me_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated or invalid token"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/drivers": {
      "get": {
        "tags": [
          "Drivers"
        ],
        "summary": "List all drivers",
        "description": "Returns all drivers.",
        "operationId": "list_drivers_drivers_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DriverResponse"
                  },
                  "type": "array",
                  "title": "Response List Drivers Drivers Get"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Drivers"
        ],
        "summary": "Create a new driver",
        "description": "Register a new driver. Requires fleet_manager role. License number must be unique.",
        "operationId": "create_driver_drivers_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DriverCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriverResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Insufficient role"
          },
          "409": {
            "description": "License number already exists"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/drivers/{driver_id}": {
      "get": {
        "tags": [
          "Drivers"
        ],
        "summary": "Get driver by ID",
        "description": "Returns a single driver by UUID.",
        "operationId": "get_driver_drivers__driver_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "driver_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Driver Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriverResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "404": {
            "description": "Driver not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Drivers"
        ],
        "summary": "Update a driver",
        "description": "Update driver fields. fleet_manager: all fields. safety_officer: compliance fields only (license_*, safety_score, status).",
        "operationId": "update_driver_drivers__driver_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "driver_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Driver Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DriverUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriverResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Insufficient role"
          },
          "404": {
            "description": "Driver not found"
          },
          "409": {
            "description": "License number already exists"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Drivers"
        ],
        "summary": "Delete a driver",
        "description": "Delete a driver by ID. Requires fleet_manager role.",
        "operationId": "delete_driver_drivers__driver_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "driver_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Driver Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Insufficient role"
          },
          "404": {
            "description": "Driver not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DriverCreate": {
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name",
            "examples": [
              "Alex Johnson"
            ]
          },
          "license_number": {
            "type": "string",
            "minLength": 1,
            "title": "License Number",
            "examples": [
              "DL-2024-0042"
            ]
          },
          "license_category": {
            "type": "string",
            "title": "License Category",
            "examples": [
              "C"
            ]
          },
          "license_expiry_date": {
            "type": "string",
            "format": "date",
            "title": "License Expiry Date",
            "examples": [
              "2026-12-31"
            ]
          },
          "contact_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Number",
            "examples": [
              "+1-555-0142"
            ]
          },
          "safety_score": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Safety Score",
            "default": 100.0,
            "examples": [
              95.0
            ]
          },
          "status": {
            "$ref": "#/components/schemas/DriverStatus",
            "default": "Available",
            "examples": [
              "Available"
            ]
          }
        },
        "type": "object",
        "required": [
          "name",
          "license_number",
          "license_category",
          "license_expiry_date"
        ],
        "title": "DriverCreate"
      },
      "DriverResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "license_number": {
            "type": "string",
            "title": "License Number"
          },
          "license_category": {
            "type": "string",
            "title": "License Category"
          },
          "license_expiry_date": {
            "type": "string",
            "format": "date",
            "title": "License Expiry Date"
          },
          "contact_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Number"
          },
          "safety_score": {
            "type": "number",
            "title": "Safety Score"
          },
          "status": {
            "$ref": "#/components/schemas/DriverStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "license_number",
          "license_category",
          "license_expiry_date",
          "contact_number",
          "safety_score",
          "status",
          "created_at"
        ],
        "title": "DriverResponse"
      },
      "DriverStatus": {
        "type": "string",
        "enum": [
          "Available",
          "On Trip",
          "Off Duty",
          "Suspended"
        ],
        "title": "DriverStatus"
      },
      "DriverUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "examples": [
              "Alex Johnson"
            ]
          },
          "license_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "License Number",
            "examples": [
              "DL-2024-0042"
            ]
          },
          "license_category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "License Category",
            "examples": [
              "C"
            ]
          },
          "license_expiry_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "License Expiry Date",
            "examples": [
              "2026-12-31"
            ]
          },
          "contact_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Number",
            "examples": [
              "+1-555-0142"
            ]
          },
          "safety_score": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Safety Score",
            "examples": [
              95.0
            ]
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DriverStatus"
              },
              {
                "type": "null"
              }
            ],
            "examples": [
              "Available"
            ]
          }
        },
        "type": "object",
        "title": "DriverUpdate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Token": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          }
        },
        "type": "object",
        "required": [
          "access_token"
        ],
        "title": "Token"
      },
...
      "UserLogin": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "examples": [
              "alice@transitops.com"
            ]
          },
          "password": {
            "type": "string",
            "title": "Password",
            "examples": [
              "SecureP@ss1"
            ]
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "UserLogin"
      },
      "UserRegister": {
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name",
            "examples": [
              "Alice Fleet"
            ]
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "examples": [
              "alice@transitops.com"
            ]
          },
          "password": {
            "type": "string",
            "minLength": 6,
            "title": "Password",
            "examples": [
              "SecureP@ss1"
            ]
          },
          "role": {
            "$ref": "#/components/schemas/UserRole",
            "examples": [
              "fleet_manager"
            ]
          }
        },
        "type": "object",
        "required": [
          "name",
          "email",
          "password",
          "role"
        ],
        "title": "UserRegister"
      },
      "UserResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "email",
          "role",
          "is_active",
          "created_at"
        ],
        "title": "UserResponse"
      },
      "UserRole": {
        "type": "string",
        "enum": [
          "fleet_manager",
          "driver",
          "safety_officer",
          "financial_analyst"
        ],
        "title": "UserRole"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}